Move relevant exceptions to module with ARC REST code
