+++ src/act/arc/aCTSubmitter.py
@@ -1,17 +1,15 @@
 import datetime
 import os
-
 from http.client import HTTPException
 from json import JSONDecodeError
 from random import shuffle
 from ssl import SSLError
 from urllib.parse import urlparse
 
-from act.arc.rest import RESTClient
-from act.common.exceptions import (ACTError, ARCHTTPError,
-                                   DescriptionParseError,
-                                   DescriptionUnparseError, InputFileError, NoValueInARCResult)
 from act.arc.aCTStatus import ARC_STATE_MAPPING
+from act.arc.rest import (ARCError, ARCHTTPError, DescriptionParseError,
+                          DescriptionUnparseError, InputFileError,
+                          NoValueInARCResult, RESTClient)
 from act.common.aCTProcess import aCTProcess
 
 
@@ -168,7 +166,7 @@ class aCTSubmitter(aCTProcess):
             try:
                 restClient = RESTClient(url.hostname, port=url.port, proxypath=proxypath)
                 jobs = restClient.submitJobs(queue, jobs, self.log)
-            except (HTTPException, ConnectionError, SSLError, ACTError, ARCHTTPError, TimeoutError) as exc:
+            except (HTTPException, ConnectionError, SSLError, ARCError, ARCHTTPError, TimeoutError) as exc:
                 self.log.error(f"Error submitting jobs to ARC: {exc}")
                 self.setJobsArcstate(jobs, "tosubmit", commit=True)
                 continue
@@ -311,7 +309,7 @@ class aCTSubmitter(aCTProcess):
             try:
                 restClient = RESTClient(url.hostname, port=url.port, proxypath=proxypath)
                 toARCKill = restClient.killJobs(toARCKill)
-            except (HTTPException, ConnectionError, SSLError, ACTError, ARCHTTPError, TimeoutError) as exc:
+            except (HTTPException, ConnectionError, SSLError, ARCError, ARCHTTPError, TimeoutError) as exc:
                 self.log.error(f"Error killing jobs in ARC: {exc}")
             except JSONDecodeError as exc:
                 self.log.error(f"Invalid JSON response from ARC: {exc}")
@@ -390,7 +388,7 @@ class aCTSubmitter(aCTProcess):
             try:
                 restClient = RESTClient(url.hostname, port=url.port, proxypath=proxypath)
                 toARCClean = restClient.cleanJobs(toARCClean)
-            except (HTTPException, ConnectionError, SSLError, ACTError, ARCHTTPError, TimeoutError) as exc:
+            except (HTTPException, ConnectionError, SSLError, ARCError, ARCHTTPError, TimeoutError) as exc:
                 self.log.error(f"Error killing jobs in ARC: {exc}")
             except JSONDecodeError as exc:
                 self.log.error(f"Invalid JSON response from ARC: {exc}")
@@ -483,7 +481,7 @@ class aCTSubmitter(aCTProcess):
                 # restart jobs
                 restClient.restartJobs(torestart)
 
-            except (HTTPException, ConnectionError, SSLError, ACTError, ARCHTTPError, TimeoutError) as exc:
+            except (HTTPException, ConnectionError, SSLError, ARCError, ARCHTTPError, TimeoutError) as exc:
                 self.log.error(f"Error killing jobs in ARC: {exc}")
             except JSONDecodeError as exc:
                 self.log.error(f"Invalid JSON response from ARC: {exc}")
